remove contracts, simplify structure
