--- src/domain/contract/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod file_system;
-pub(crate) mod minecraft_mods_service;
