--- src/domain/entities/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod minecraft_mod;
