+++ src/lib.rs
@@ -2,13 +2,11 @@ use clap::Parser;
 
 use cli::{Cli, Commands};
 
-mod adapter;
 mod cli;
 mod domain;
 mod infrastructure;
+mod presentation;
 mod service;
-mod utils;
-mod view;
 
 pub async fn run() {
     let args = Cli::parse();
@@ -17,16 +15,16 @@ pub async fn run() {
         Commands::Add {
             minecraft_version,
             id_or_slug,
-        } => view::add::add_view(&minecraft_version, &id_or_slug).await,
+        } => service::minecraft_mods::add(&minecraft_version, &id_or_slug).await,
         Commands::Rm {
             minecraft_file: minecraft_mod_file_name,
-        } => view::rm::rm_view(&minecraft_mod_file_name),
-        Commands::List => view::list::list_view(),
+        } => service::minecraft_mods::rm(&minecraft_mod_file_name),
+        Commands::List => service::minecraft_mods::list(),
         Commands::Latest { minecraft_version } => {
-            view::latest::latest_view(&minecraft_version).await
+            service::minecraft_mods::latest(&minecraft_version).await
         }
         Commands::Support { minecraft_version } => {
-            view::support::support_view(&minecraft_version).await
+            service::minecraft_mods::support(&minecraft_version).await
         }
     }
 }
