+++ src/presentation/view/add.rs
@@ -1,0 +1,9 @@
+use colored::Colorize;
+
+pub(crate) fn opening_message(id_or_slug: &str) {
+    println!(
+        "\r\n{} {id_or_slug} {}",
+        "Searching".bright_cyan(),
+        "on Modrinth...".bright_cyan()
+    );
+}
