+++ src/presentation/view/list.rs
@@ -1,0 +1,8 @@
+use colored::Colorize;
+
+pub(crate) fn opening_message() {
+    println!(
+        "{}...",
+        "Displaying Minecraft mods in the mods directory".bright_cyan()
+    );
+}
