+++ src/presentation/view/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod latest;
 pub(crate) mod list;
 pub(crate) mod rm;
 pub(crate) mod support;
+pub(crate) mod update;
